修复销毁窗口造成程序崩溃的bug

+++ Src/TcpServer/SAKTcpServerDebugPage.cc
@@ -33,8 +33,11 @@ SAKTcpServerDebugPage::SAKTcpServerDebugPage(QWidget *parent)
 SAKTcpServerDebugPage::~SAKTcpServerDebugPage()
 {
     tcpServerDeviceController->deleteLater();
-    tcpServerDevice->terminate();
-    delete tcpServerDevice;
+
+    if (tcpServerDevice){
+        tcpServerDevice->terminate();
+        delete tcpServerDevice;
+    }
 }
 
 void SAKTcpServerDebugPage::setUiEnable(bool enable)
